Update README.md
MM added text from a GoogleDoc that was in progress and outlined more specific protocols. It seemed that much of the content was duplicated between the repository README.md and what was called 'specifications.md', so MM decided that the simpler thing to do would be to have all instructions in one place. This decision can, of course, be changed and a more specific instructions document can be created again.