Update README.md
Editing from #3 in the task list. I removed the point about keeping track of behaviour data because Melina and I discussed that BAM is not a warehouse for information other than abundance data that is used in WildTrax. 

Added how data goes from 'not started' to 'in progress' in the WT-Status column of the proejct_integration GoogleSheet.